Implemented sign change options for PLS. Can you please add a comments to the summarize print which sign change option is used.
For lineare models, I compraed the results to my expectation and they matched. For the seoncd-order model is more difficult. It runs through but it is not tested yet.

Probably it can be done more efficiently.

the lapply needs to be replaced by the future apply and the seed needss to be removed.